@@ -166,7 +166,7 @@ namespace ExifTimeShift.Views.Controls
 			}
 			finally
 			{
-				if (fod != null)
+				if (fod is not null)
 					Marshal.FinalReleaseComObject(fod);
 			}
 		}
